Revert "Merge branch 'feature/federation' into feature/federation-downloader"
This reverts commit a1ccc53dd8c824a1f35d5e33d40bfdab7de6c625.
