+++ common/federation/pom.xml
@@ -3,23 +3,40 @@
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <parent>
-    <artifactId>common</artifactId>
-    <groupId>org.opencwa</groupId>
-    <version>${revision}</version>
-    <relativePath>../pom.xml</relativePath>
+    <groupId>org.springframework.boot</groupId>
+    <artifactId>spring-boot-starter-parent</artifactId>
+    <version>2.3.1.RELEASE</version>
+    <relativePath></relativePath>
   </parent>
   <modelVersion>4.0.0</modelVersion>
 
+  <groupId>org.opencwa</groupId>
   <artifactId>federation</artifactId>
+  <version>${revision}</version>
+
+  <organization>
+    <name>SAP SE</name>
+  </organization>
+
+  <properties>
+    <java.version>11</java.version>
+    <maven.compiler.source>11</maven.compiler.source>
+    <maven.compiler.target>11</maven.compiler.target>
+    <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+    <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
+    <sonar.projectKey>corona-warn-app_cwa-server_common_federation</sonar.projectKey>
+  </properties>
 
   <dependencies>
     <dependency>
       <groupId>org.springframework.cloud</groupId>
       <artifactId>spring-cloud-starter-openfeign</artifactId>
+      <version>2.2.3.RELEASE</version>
     </dependency>
     <dependency>
       <groupId>io.github.openfeign</groupId>
       <artifactId>feign-httpclient</artifactId>
+      <version>11.0</version>
     </dependency>
     <dependency>
       <groupId>org.springframework.boot</groupId>
@@ -47,24 +64,33 @@
       <artifactId>spring-boot-starter-log4j2</artifactId>
     </dependency>
     <dependency>
+      <!-- https://nvd.nist.gov/vuln/detail/CVE-2020-9488 -->
       <groupId>org.apache.logging.log4j</groupId>
       <artifactId>log4j-core</artifactId>
+      <version>2.13.3</version>
     </dependency>
     <dependency>
+      <!-- https://nvd.nist.gov/vuln/detail/CVE-2020-9488 -->
       <groupId>org.apache.logging.log4j</groupId>
       <artifactId>log4j-api</artifactId>
+      <version>2.13.3</version>
     </dependency>
     <dependency>
+      <!-- Due to https://nvd.nist.gov/vuln/detail/CVE-2020-10683 -->
       <groupId>org.dom4j</groupId>
       <artifactId>dom4j</artifactId>
+      <version>2.1.3</version>
     </dependency>
     <dependency>
+      <!-- https://nvd.nist.gov/vuln/detail/CVE-2017-18640 -->
       <groupId>org.yaml</groupId>
       <artifactId>snakeyaml</artifactId>
+      <version>1.26</version>
     </dependency>
     <dependency>
       <groupId>io.github.openfeign</groupId>
       <artifactId>feign-core</artifactId>
+      <version>11.0</version>
     </dependency>
   </dependencies>
 
@@ -95,14 +121,17 @@
       <plugin>
         <groupId>org.apache.maven.plugins</groupId>
         <artifactId>maven-site-plugin</artifactId>
+        <version>3.7.1</version>
       </plugin>
       <plugin>
         <groupId>org.apache.maven.plugins</groupId>
         <artifactId>maven-project-info-reports-plugin</artifactId>
+        <version>3.0.0</version>
       </plugin>
       <plugin>
         <groupId>org.codehaus.mojo</groupId>
         <artifactId>license-maven-plugin</artifactId>
+        <version>2.0.0</version>
         <configuration>
           <includes>**/*.java</includes>
           <copyrightOwners>${project.organization.name} and all other contributors</copyrightOwners>
@@ -118,6 +147,7 @@
         <!-- see https://maven.apache.org/maven-ci-friendly.html#install-deploy -->
         <groupId>org.codehaus.mojo</groupId>
         <artifactId>flatten-maven-plugin</artifactId>
+        <version>1.1.0</version>
         <configuration>
           <updatePomFile>true</updatePomFile>
           <flattenMode>resolveCiFriendliesOnly</flattenMode>
@@ -142,6 +172,7 @@
       <plugin>
         <groupId>org.apache.maven.plugins</groupId>
         <artifactId>maven-checkstyle-plugin</artifactId>
+        <version>3.1.1</version>
         <configuration>
           <configLocation>../../codestyle/checkstyle.xml</configLocation>
           <excludes>**/module-info.java,**/target/**/*</excludes>
