+++ common/federation/src/main/java/app/coronawarn/server/common/federation/client/FederationGatewayClient.java
@@ -24,7 +24,6 @@ import feign.Response;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestHeader;
 
 /**
@@ -49,23 +48,4 @@ public interface FederationGatewayClient {
       headers = {"Accept=application/protobuf; version=1.0", "X-SSL-Client-SHA256=abcd", "X-SSL-Client-DN=C=PL"})
   Response getDiagnosisKeys(@RequestHeader("batchTag") String batchTag,
       @PathVariable("date") String date);
-
-  /**
-   * HTTP POST request federation gateway endpoint /diagnosiskyes/upload.
-   * @param raw Payload body. This property contains a raw byte array with the encoded protobuf DiagnosisKeyBatch.
-   * @param accept HTTP Header Accept.
-   * @param shaClient HTTP Header X-SSL-Client-SHA256.
-   * @param dnClient HTTP Header X-SSL-Client-DN.
-   * @param batchTag Unique batchTag to be identified by EFGS.
-   * @param batchSignature Batch Signature as per PKCS#7 spec using Authorized Signing Certificate.
-   */
-  @PostMapping(value = "/diagnosiskeys/upload",
-      consumes = "application/protobuf; version=1.0")
-  String postBatchUpload(
-      byte[] raw,
-      @RequestHeader("Accept") String accept,
-      @RequestHeader("X-SSL-Client-SHA256") String shaClient,
-      @RequestHeader("X-SSL-Client-DN") String dnClient,
-      @RequestHeader("batchTag") String batchTag,
-      @RequestHeader("batchSignature") String batchSignature);
 }
