+++ common/persistence/src/main/java/app/coronawarn/server/common/persistence/service/DiagnosisKeyService.java
@@ -26,13 +26,13 @@ import static org.springframework.data.util.StreamUtils.createStreamFromIterator
 
 import app.coronawarn.server.common.persistence.domain.DiagnosisKey;
 import app.coronawarn.server.common.persistence.repository.DiagnosisKeyRepository;
-import app.coronawarn.server.common.persistence.service.common.ValidDiagnosisKeyFilter;
 import io.micrometer.core.annotation.Timed;
 import java.time.Instant;
 import java.time.LocalDateTime;
 import java.util.Collection;
 import java.util.List;
 import java.util.stream.Collectors;
+import javax.validation.ConstraintViolation;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.data.domain.Sort;
@@ -45,11 +45,9 @@ public class DiagnosisKeyService {
 
   private static final Logger logger = LoggerFactory.getLogger(DiagnosisKeyService.class);
   private final DiagnosisKeyRepository keyRepository;
-  private final ValidDiagnosisKeyFilter validationFilter;
 
-  public DiagnosisKeyService(DiagnosisKeyRepository keyRepository, ValidDiagnosisKeyFilter filter) {
+  public DiagnosisKeyService(DiagnosisKeyRepository keyRepository) {
     this.keyRepository = keyRepository;
-    this.validationFilter = filter;
   }
 
   /**
@@ -78,7 +76,7 @@ public class DiagnosisKeyService {
   public List<DiagnosisKey> getDiagnosisKeys() {
     List<DiagnosisKey> diagnosisKeys = createStreamFromIterator(
         keyRepository.findAll(Sort.by(Direction.ASC, "submissionTimestamp")).iterator()).collect(Collectors.toList());
-    return validationFilter.filter(diagnosisKeys);
+    return this.filterValidDiagnosisKeys(diagnosisKeys);
   }
 
   /**
@@ -91,7 +89,31 @@ public class DiagnosisKeyService {
   public List<DiagnosisKey> getDiagnosisKeysByVisitedCountry(String countryCode) {
     var diagnosisKeys = createStreamFromIterator(
         keyRepository.findAllKeysWhereVisitedCountryContains(countryCode).iterator()).collect(Collectors.toList());
-    return validationFilter.filter(diagnosisKeys);
+    return this.filterValidDiagnosisKeys(diagnosisKeys);
+  }
+
+  private List<DiagnosisKey> filterValidDiagnosisKeys(List<DiagnosisKey> diagnosisKeys) {
+    List<DiagnosisKey> validDiagnosisKeys =
+        diagnosisKeys.stream().filter(DiagnosisKeyService::isDiagnosisKeyValid).collect(Collectors.toList());
+
+    int numberOfDiscardedKeys = diagnosisKeys.size() - validDiagnosisKeys.size();
+    logger.info("Retrieved {} diagnosis key(s). Discarded {} diagnosis key(s) from the result as invalid.",
+        diagnosisKeys.size(), numberOfDiscardedKeys);
+
+    return validDiagnosisKeys;
+  }
+
+  private static boolean isDiagnosisKeyValid(DiagnosisKey diagnosisKey) {
+    Collection<ConstraintViolation<DiagnosisKey>> violations = diagnosisKey.validate();
+    boolean isValid = violations.isEmpty();
+
+    if (!isValid) {
+      List<String> violationMessages =
+          violations.stream().map(ConstraintViolation::getMessage).collect(Collectors.toList());
+      logger.warn("Validation failed for diagnosis key from database. Violations: {}", violationMessages);
+    }
+
+    return isValid;
   }
 
   /**
