--- docker-compose-test-secrets/efgs_signing_cert.pem
@@ -1,12 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIBwzCCAWigAwIBAgIJANdK6mL/gUWiMAoGCCqGSM49BAMCMD8xHTAbBgNVBAMM
-FENXQSBUZXN0IENlcnRpZmljYXRlMREwDwYDVQQLDAhDV0EtVGVhbTELMAkGA1UE
-BhMCREUwHhcNMjAwOTAxMTIyOTQxWhcNMjEwOTAxMTIyOTQxWjA/MR0wGwYDVQQD
-DBRDV0EgVGVzdCBDZXJ0aWZpY2F0ZTERMA8GA1UECwwIQ1dBLVRlYW0xCzAJBgNV
-BAYTAkRFMFkwEwYHKoZIzj0CAQYIKoZIzj0DAQcDQgAEosYj9mhDbr4jSuI3kWLY
-sr70Iq8FPGLBjeoW2SzBNo4NIjDSNalTLKiRfyXCojmUJC+ZhZnQA7DXZADD/WNX
-XaNNMEswMQYDVR0lBCowKAYIKwYBBQUHAwEGCCsGAQUFBwMCBggrBgEFBQcDAwYI
-KwYBBQUHAwQwCQYDVR0TBAIwADALBgNVHQ8EBAMCBeAwCgYIKoZIzj0EAwIDSQAw
-RgIhAPjFZGoeamLmENf7O5hBzJwbmIt8btKfexZPwZvGKZv4AiEA3Wp8474SKlx7
-F0TLpKfdA6fs3Ri9rNGiJOQ6IZmu20s=
------END CERTIFICATE-----
