--- docker-compose-test-secrets/efgs_signing_key.pem
@@ -1,5 +1,0 @@
------BEGIN EC PRIVATE KEY-----
-MHcCAQEEIPDeu93N7BDruxwyDs3uV1oWQ8pewXMSoY0B5hMK9vpfoAoGCCqGSM49
-AwEHoUQDQgAEosYj9mhDbr4jSuI3kWLYsr70Iq8FPGLBjeoW2SzBNo4NIjDSNalT
-LKiRfyXCojmUJC+ZhZnQA7DXZADD/WNXXQ==
------END EC PRIVATE KEY-----
