--- docker-compose-test-secrets/efgs_x509_thumbprint.txt
@@ -1,1 +1,0 @@
-8258910F8C6B46568C2ECBF040B2F65BDC770E78C09E58776EE92F4FA9D2AA57
