+++ services/upload/src/main/java/app/coronawarn/server/services/federation/upload/Application.java
@@ -20,7 +20,6 @@
 
 package app.coronawarn.server.services.federation.upload;
 
-import app.coronawarn.server.services.federation.upload.config.UploadServiceConfig;
 import java.util.Arrays;
 import java.util.List;
 import org.apache.logging.log4j.LogManager;
@@ -31,24 +30,21 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.domain.EntityScan;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import org.springframework.context.ApplicationContext;
 import org.springframework.context.EnvironmentAware;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.core.env.Environment;
 import org.springframework.data.jdbc.repository.config.EnableJdbcRepositories;
 
 /**
- * Service responsible for creating batches of diagnosis keys and uploading them to the Federation Gateway conforming to
- * the EU specification. Its source of data is a dedicated table where diagnosis keys are replicated during the
- * submission process.
+ * Service responsible for creating batches of diagnosis keys and uploading them to the Federation Gateway
+ * conforming to the EU specification. Its source of data is a dedicated table where diagnosis keys are replicated
+ * during the submission process.
  */
 @SpringBootApplication
 @EnableJdbcRepositories(basePackages = "app.coronawarn.server.common.persistence")
 @EntityScan(basePackages = "app.coronawarn.server.common.persistence")
-@ComponentScan({"app.coronawarn.server.common.persistence",
-    "app.coronawarn.server.services.federation.upload",
-    "app.coronawarn.server.common.federation.client"})
-@EnableConfigurationProperties({UploadServiceConfig.class})
+@ComponentScan({"app.coronawarn.server.common.persistence", "app.coronawarn.server.services.federation.upload"})
+@EnableConfigurationProperties
 public class Application implements EnvironmentAware, DisposableBean {
 
   private static final Logger logger = LoggerFactory.getLogger(Application.class);
@@ -61,15 +57,6 @@ public class Application implements EnvironmentAware, DisposableBean {
   public void destroy() {
     logger.info("Shutting down log4j2.");
     LogManager.shutdown();
-  }
-
-  /**
-   * Terminates this application with exit code 1 (general error).
-   */
-  public static void killApplication(ApplicationContext appContext) {
-    SpringApplication.exit(appContext);
-    logger.error("Federation Upload Service terminated abnormally.");
-    System.exit(1);
   }
 
   @Override
