--- services/upload/src/main/java/app/coronawarn/server/services/federation/upload/client/FederationUploadClient.java
@@ -1,9 +1,0 @@
-package app.coronawarn.server.services.federation.upload.client;
-
-import app.coronawarn.server.services.federation.upload.payload.UploadPayload;
-
-public interface FederationUploadClient {
-
-  void postBatchUpload(UploadPayload uploadPayload);
-
-}
