--- services/upload/src/main/java/app/coronawarn/server/services/federation/upload/keys/DiagnosisKeyLoader.java
@@ -1,11 +1,0 @@
-package app.coronawarn.server.services.federation.upload.keys;
-
-
-import app.coronawarn.server.common.persistence.domain.DiagnosisKey;
-import java.util.List;
-
-public interface DiagnosisKeyLoader {
-
-  List<DiagnosisKey> loadDiagnosisKeys();
-
-}
