--- services/upload/src/test/resources/testprivatekey.pem
@@ -1,5 +1,0 @@
------BEGIN PRIVATE KEY-----
-MIGHAgEAMBMGByqGSM49AgEGCCqGSM49AwEHBG0wawIBAQQgjPus/CYkGPQhgBtC
-MtG+vC2BQkMDkbcswEP182fFJ7ehRANCAATWXHdqcXz9bNmoxoLlCK8wHpy/ws28
-CyvzB3xgOEeBBbuGKFPBcu5PRxN1mLe3jU0K2bvcLDcvbuP5/Qfq25Z3
------END PRIVATE KEY-----
