--- services/upload/src/test/resources/testpublickey.pem
@@ -1,12 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIBtjCCAVygAwIBAgIJAKHUcVXhuN3dMAoGCCqGSM49BAMCMDkxFDASBgNVBAMM
-C0ZHUyBHZXJtYW55MRQwEgYDVQQLDAtDb3JvbmEtVGVhbTELMAkGA1UEBhMCREUw
-HhcNMjAwODI4MTIzODEwWhcNMjEwODI4MTIzODEwWjA5MRQwEgYDVQQDDAtGR1Mg
-R2VybWFueTEUMBIGA1UECwwLQ29yb25hLVRlYW0xCzAJBgNVBAYTAkRFMFkwEwYH
-KoZIzj0CAQYIKoZIzj0DAQcDQgAE1lx3anF8/WzZqMaC5QivMB6cv8LNvAsr8wd8
-YDhHgQW7hihTwXLuT0cTdZi3t41NCtm73Cw3L27j+f0H6tuWd6NNMEswMQYDVR0l
-BCowKAYIKwYBBQUHAwEGCCsGAQUFBwMCBggrBgEFBQcDAwYIKwYBBQUHAwQwCQYD
-VR0TBAIwADALBgNVHQ8EBAMCBeAwCgYIKoZIzj0EAwIDSAAwRQIhAOtEEWyA4eG2
-GopTSvi+yXhP/EuyYT4dDCoCdCVVQurvAiA/t1YHxWtVQD58E1jIaOK6+mkrECxx
-vTq/go9KgBdn3g==
------END CERTIFICATE-----
